[fix] RouletteMemberコンポーネントを作成。「話題をチェンジ」で発生するバグを修正 Co-authored-by: sabarasaraba <sabarasaraba@users.noreply.github.com> Co-authored-by: ishuuuu <ishuuuu@users.noreply.github.com>
