+++ src/context/RouletteContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const RouletteContext = createContext();
+
+export default RouletteContext;
